Fix cecil analysis and fix CallExtensionMethodAsExtensionAnalyzer
Cecil analysis was not working properly after Roslyn RTM, as the message
changed. Now we are correctly inspecting for the message, which involves
looking for an AD0001 diagnostic on the build log.
After reenabling the cecil build a bug was found on the
CallExtensionMethodAsExtensionAnalyzer, so that was also fixed.
